requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
